Remove obsolete shared model stub
