+++ app/models/__init__.py
@@ -1,1 +1,4 @@
 from .division import Division  # noqa: F401
+from .rank import Rank  # noqa: F401
+from .basho import Basho  # noqa: F401
+from .rikishi import Heya, Rikishi, Shusshin  # noqa: F401
